Fixed the issue with textures not loading on 64bit linux.
On 64bit linux, the long type is 64bits long, while one 32bit linux and both 64 and 32bit windows it is only 32bits long.
There are probally many other places in the code where longs are causing issues.

Right now, the menu runs fine, but it crashes when I try to start new a game.